PHP to HTML
The index file doesn’t use any php so it should be an HTML file. this
makes it also possible to display the mirror without a server. NOTE: If
you do run the mirror on a server, it wil NOT display the calendar, nor
wil it automatically refresh the page when the config file is adjusted.